# TP7_API_LARAVEL_INFYOM/storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel Generator APIs",
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/compteepsxes": {
            "get": {
                "tags": [
                    "Compteepsx"
                ],
                "summary": "Get a listing of the Compteepsxes.",
                "description": "Get all Compteepsxes",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Compteepsx"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Compteepsx"
                ],
                "summary": "Store a newly created Compteepsx in storage",
                "description": "Store Compteepsx",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Compteepsx that should be stored",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Compteepsx"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Compteepsx"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/compteepsxes/{id}": {
            "get": {
                "tags": [
                    "Compteepsx"
                ],
                "summary": "Display the specified Compteepsx",
                "description": "Get Compteepsx",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Compteepsx",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Compteepsx"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Compteepsx"
                ],
                "summary": "Update the specified Compteepsx in storage",
                "description": "Update Compteepsx",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Compteepsx",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Compteepsx that should be updated",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Compteepsx"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Compteepsx"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Compteepsx"
                ],
                "summary": "Remove the specified Compteepsx from storage",
                "description": "Delete Compteepsx",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Compteepsx",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Compteepsx": {
            "required": [
                "accountNumber",
                "cleRIB",
                "solde",
                "dateCreation",
                "nextRemunDate",
                "hostAgency"
            ],
            "properties": {
                "id": {
                    "description": "id",
                    "type": "integer",
                    "format": "int32"
                },
                "owner_id": {
                    "description": "owner_id",
                    "type": "integer",
                    "format": "int32"
                },
                "accountNumber": {
                    "description": "accountNumber",
                    "type": "string"
                },
                "cleRIB": {
                    "description": "cleRIB",
                    "type": "integer",
                    "format": "int32"
                },
                "solde": {
                    "description": "solde",
                    "type": "number",
                    "format": "number"
                },
                "dateCreation": {
                    "description": "dateCreation",
                    "type": "string"
                },
                "activeDate": {
                    "description": "activeDate",
                    "type": "string"
                },
                "nextRemunDate": {
                    "description": "nextRemunDate",
                    "type": "string"
                },
                "closeDate": {
                    "description": "closeDate",
                    "type": "string"
                },
                "idUser": {
                    "description": "idUser",
                    "type": "integer",
                    "format": "int32"
                },
                "hostAgency": {
                    "description": "hostAgency",
                    "type": "integer",
                    "format": "int32"
                },
                "idOpeningFees": {
                    "description": "idOpeningFees",
                    "type": "integer",
                    "format": "int32"
                },
                "created_at": {
                    "description": "created_at",
                    "type": "string",
                    "format": "date-time"
                },
                "updated_at": {
                    "description": "updated_at",
                    "type": "string",
                    "format": "date-time"
                },
                "deleted_at": {
                    "description": "deleted_at",
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
...
    }
}